refactor: move metadata service into new file
